feat: stomp config 수정
