--- src/main/java/com/hack/stock2u/chat/web/StompWepSocketConfig.java
@@ -1,35 +1,0 @@
-package com.hack.stock2u.chat.web;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.event.EventListener;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.session.web.socket.events.SessionConnectEvent;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-import org.springframework.web.socket.config.annotation.WebSocketTransportRegistration;
-import org.springframework.web.socket.messaging.SessionDisconnectEvent;
-
-@Configuration
-@EnableWebSocketMessageBroker
-public class StompWepSocketConfig implements WebSocketMessageBrokerConfigurer {
-
-  //stomphandler만들기
-  //웹소켓 핸드셰이크 커넥션을 생성할 경로
-  @Override
-  public void registerStompEndpoints(StompEndpointRegistry registry) {
-
-    registry.addEndpoint("/ws").setAllowedOriginPatterns("*").withSockJS();
-  }
-
-  @Override
-  public void configureMessageBroker(MessageBrokerRegistry registry) {
-
-    registry.setApplicationDestinationPrefixes("/api");  //
-    // @MessageMapping 메서드로 라우팅된다.//pub
-    registry.enableSimpleBroker("/topic", "queue");
-    // SimpleBroker는 해당하는 경로를 SUBSCRIBE하는 Client에게 메세지를 전달하는 간단한 작업을 수행
-    //sub
-  }
-}
