+++ src/main/resources/templates/chat.html
@@ -1,0 +1,57 @@
+<!doctype html>
+<html xmlns:th="http://www.thymeleaf.org">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport"
+              content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+        <meta http-equiv="X-UA-Compatible" content="ie=edge">
+        <title>Document</title>
+    </head>
+    <body>
+        <script src="https://cdn.jsdelivr.net/npm/@stomp/stompjs@5.0.0/bundles/stomp.umd.min.js"></script>
+        <div id="main-content" class="container">
+            <div class="row">
+                <div class="col-md-6">
+                    <form class="form-inline">
+                        <div class="form-group">
+                            <label for="connect">WebSocket connection:</label>
+                            <button id="connect" class="btn btn-default" type="submit">Connect</button>
+                            <button id="disconnect" class="btn btn-default" type="submit" disabled="disabled">Disconnect
+                            </button>
+                        </div>
+                    </form>
+                </div>
+                <div class="col-md-6">
+                    <form class="form-inline">
+                        <div class="form-group">
+                            <label for="name">What is your name?</label>
+                            <input type="text" id="name" class="form-control" placeholder="Your name here...">
+                        </div>
+                        <button id="send" class="btn btn-default" type="submit">Send</button>
+                    </form>
+                </div>
+            </div>
+            <div class="row">
+                <div class="col-md-12">
+                    <table id="conversation" class="table table-striped">
+                        <thead>
+                        <tr>
+                            <th>Greetings</th>
+                        </tr>
+                        </thead>
+                        <tbody id="greetings">
+                        </tbody>
+                    </table>
+                </div>
+            </div>
+        </div>
+        <script>
+            const client = new StompJs.Client({
+                brokerURL: 'ws://localhost:8081/api/ws'
+            });
+
+            client.activate();
+            console.log(client);
+        </script>
+    </body>
+</html>
